requirements.txt
tornado
wtforms
peewee
pymysql
aiomysql
pyJWT
aiofiles
